added parsing to more efficiently update userlist
ReaderThread now parses user joining and leaving messages to update userlist. This does mean that user joining and leaving messages will have to be standardized in the RFC for this to work on all servers.

updateUserList now no longer asks the server for a new list, but the updateThread still calls updateUserList on a 5 second interval.